my-weed-app/api/requirements.txt
tflite-runtime
numpy
pillow
requests